refactor widget to show name of numbersequence refactor registry to reduce search results add primes as a number sequence bug fix in logic - find number refactor logic add tests
