+++ lib/application/category_views/selector_lists/number_sequences/numbersequence_primes_selection.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+import 'package:gc_wizard/application/registry.dart';
+import 'package:gc_wizard/common_widgets/gcw_selection.dart';
+import 'package:gc_wizard/common_widgets/gcw_tool.dart';
+import 'package:gc_wizard/common_widgets/gcw_toollist.dart';
+import 'package:gc_wizard/tools/science_and_technology/number_sequences/primes/widget/primes.dart';
+import 'package:gc_wizard/utils/ui_dependent_utils/common_widget_utils.dart';
+
+class NumberSequencePrimesSelection extends GCWSelection {
+  const NumberSequencePrimesSelection({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final List<GCWTool> _toolList = registeredTools.where((element) {
+      return [
+        className(const NumberSequencePrimesNthNumber()),
+        className(const NumberSequencePrimesRange()),
+        className(const NumberSequencePrimesDigits()),
+        className(const NumberSequencePrimesCheckNumber()),
+        className(const NumberSequencePrimesContainsDigits()),
+      ].contains(className(element.tool));
+    }).toList();
+
+    return GCWToolList(toolList: _toolList);
+  }
+}
