+++ lib/application/category_views/selector_lists/number_sequences/numbersequence_selection.dart
@@ -20,6 +20,7 @@ import 'package:gc_wizard/application/category_views/selector_lists/number_seque
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_perfectnumbers_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_permutableprimes_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_primarypseudoperfectnumbers_selection.dart';
+import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_primes_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_recaman_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_sublimenumbers_selection.dart';
 import 'package:gc_wizard/application/category_views/selector_lists/number_sequences/numbersequence_superperfectnumbers_selection.dart';
@@ -39,6 +40,7 @@ class NumberSequenceSelection extends GCWSelection {
       return [
         className(const NumberSequenceFactorialSelection()),
         className(const NumberSequenceFibonacciSelection()),
+        className(const NumberSequencePrimesSelection()),
         className(const NumberSequenceMersenneSelection()),
         className(const NumberSequenceMersennePrimesSelection()),
         className(const NumberSequenceMersenneExponentsSelection()),
