+++ lib/application/i18n/assets/en.json
@@ -3210,6 +3210,9 @@
   "numbersequence_happynumbers_title": "Happy Numbers",
   "numbersequence_happynumbers_description": "OEIS A007770, until n = 142",
   "numbersequence_happynumbers_example": "1, 7, 10, 13, 19, 23, 28, 31, 32, 44, ...",
+  "numbersequence_primes_title": "Prime Numbers",
+  "numbersequence_primes_description": "OEIS A000040, until n = 10000000",
+  "numbersequence_primes_example": "2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, ...",
   "numbersequence_nth_title": "Get n-th Number",
   "numbersequence_nth_description": "Attention: First number is n = 0, whereas some sources start at n = 1",
   "numbersequence_range_title": "Range of numbers",
