+++ lib/tools/science_and_technology/number_sequences/_common/widget/numbersequences_nthnumber.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:gc_wizard/application/i18n/logic/app_localizations.dart';
+import 'package:gc_wizard/common_widgets/dividers/gcw_text_divider.dart';
 import 'package:gc_wizard/common_widgets/outputs/gcw_default_output.dart';
 import 'package:gc_wizard/common_widgets/spinners/gcw_integer_spinner.dart';
 import 'package:gc_wizard/tools/science_and_technology/number_sequences/_common/logic/number_sequence.dart';
@@ -30,6 +31,10 @@ class _NumberSequenceNthNumberState extends State<NumberSequenceNthNumber> {
   Widget build(BuildContext context) {
     return Column(
       children: <Widget>[
+        GCWTextDivider(
+          text: i18n(context, NUMBERSEQUENCE_TITLE[widget.mode]!),
+          style: const TextStyle(fontSize: 20),
+        ),
         GCWIntegerSpinner(
           title: i18n(context, 'numbersequence_inputn'),
           min: 0,
@@ -47,6 +52,6 @@ class _NumberSequenceNthNumberState extends State<NumberSequenceNthNumber> {
   }
 
   Widget _buildOutput() {
-    return GCWDefaultOutput(child: getNumberAt(widget.mode, _currentInputN).toString());
+    return GCWDefaultOutput(child: numberSequencesGetNumberAt(widget.mode, _currentInputN).toString());
   }
 }
