+++ test/tools/science_and_technology/number_sequence/_common/logic/number_sequence_test.dart
@@ -17,11 +17,12 @@ void main() {
       {'sequence' : NumberSequencesMode.FERMAT,           'position' :  5, 'expectedOutput' : BigInt.from(4294967297)},
       {'sequence' : NumberSequencesMode.FACTORIAL,        'position' : 20, 'expectedOutput' : BigInt.from(2432902008176640000)},
       {'sequence' : NumberSequencesMode.LYCHREL,          'position' : 20, 'expectedOutput' : BigInt.from(1767)},
+      {'sequence' : NumberSequencesMode.PRIMES,           'position' : 1818, 'expectedOutput' : BigInt.from(15601)},
     ];
 
     for (var elem in _inputsToExpected) {
       test('sequence: ${elem['sequence']}, position: ${elem['position']}', () {
-        var _actual = getNumberAt(elem['sequence'] as NumberSequencesMode, elem['position'] as int?);
+        var _actual = numberSequencesGetNumberAt(elem['sequence'] as NumberSequencesMode, elem['position'] as int?);
         expect(_actual, elem['expectedOutput']);
       });
     }
@@ -42,11 +43,12 @@ void main() {
       {'sequence' : NumberSequencesMode.FERMAT,           'maxIndex' : 100, 'expectedOutput' :  5, 'number' : BigInt.from(4294967297)},
       {'sequence' : NumberSequencesMode.FACTORIAL,        'maxIndex' : 100, 'expectedOutput' : 20, 'number' : BigInt.from(2432902008176640000)},
       {'sequence' : NumberSequencesMode.LYCHREL,          'maxIndex' : 100, 'expectedOutput' : 20, 'number' : BigInt.from(1767)},
+      {'sequence' : NumberSequencesMode.PRIMES,          'maxIndex' : 10000, 'expectedOutput' : 1818, 'number' : BigInt.from(15601)},
     ];
 
     for (var elem in _inputsToExpected) {
       test('sequence: ${elem['sequence']}, number: ${elem['number']}', () {
-        var _actual = checkNumber(elem['sequence'] as NumberSequencesMode, elem['number'] as BigInt?, elem['maxIndex'] as int);
+        var _actual = numberSequencesCheckNumber(elem['sequence'] as NumberSequencesMode, elem['number'] as BigInt?, elem['maxIndex'] as int);
         expect(_actual, elem['expectedOutput']);
       });
     }
@@ -67,11 +69,12 @@ void main() {
       {'sequence' : NumberSequencesMode.FERMAT,           'start' :  3, 'stop' :  5, 'expectedOutput' : [257, 65537, 4294967297]},
       {'sequence' : NumberSequencesMode.FACTORIAL,        'start' : 10, 'stop' : 15, 'expectedOutput' : [3628800, 39916800, 479001600, 6227020800, 87178291200, 1307674368000]},
       {'sequence' : NumberSequencesMode.LYCHREL,          'start' : 10, 'stop' : 15, 'expectedOutput' : [887, 978, 986, 1495, 1497]},
+      {'sequence' : NumberSequencesMode.PRIMES,           'start' : 10, 'stop' : 15, 'expectedOutput' : [31, 37, 41, 43, 47, 53]},
     ];
 
     for (var elem in _inputsToExpected) {
       test('sequence: ${elem['sequence']}, start: ${elem['start']}, stop: ${elem['stop']}', () {
-        var _actual = getNumbersInRange(elem['sequence'] as NumberSequencesMode, elem['start'] as int?, elem['stop'] as int?);
+        var _actual = numberSequencesGetNumbersInRange(elem['sequence'] as NumberSequencesMode, elem['start'] as int?, elem['stop'] as int?);
         var length = (elem['expectedOutput'] as List<int>).length;
         for (int i = 0; i < length; i++) {
           expect(_actual[i], BigInt.from((elem['expectedOutput'] as List<int>)[i]));
@@ -95,11 +98,12 @@ void main() {
       {'sequence' : NumberSequencesMode.PELL_LUCAS,       'digits' : 4, 'expectedOutput' : [1154,2786,6726]},
       {'sequence' : NumberSequencesMode.FACTORIAL,        'digits' : 3, 'expectedOutput' : [120,720]},
       {'sequence' : NumberSequencesMode.LYCHREL,          'digits' : 3, 'expectedOutput' : [196, 295, 394, 493, 592, 689, 691, 788, 790, 879, 887, 978, 986]}, // Mark test
+      {'sequence' : NumberSequencesMode.PRIMES,          'digits' : 1, 'expectedOutput' : [2, 3, 5, 7]},
     ];
 
     for (var elem in _inputsToExpected) {
       test('sequence: ${elem['sequence']}, digits: ${elem['digits']}', () {
-        var _actual = getNumbersWithNDigits(elem['sequence'] as NumberSequencesMode, elem['digits'] as int?);
+        var _actual = numberSequencesGetNumbersWithNDigits(elem['sequence'] as NumberSequencesMode, elem['digits'] as int?);
         var length = (elem['expectedOutput'] as List<Object>).length;
         for (int i = 0; i < length; i++) {
           if (_actual[i].runtimeType.toString() == 'String') {
@@ -140,11 +144,13 @@ void main() {
           549, 1118)},
       {'sequence' : NumberSequencesMode.LYCHREL,          'maxIndex' : 1000,  'number' : '996', 'expectedOutput' : PositionOfSequenceOutput('2996',
           38, 2)},
+      {'sequence' : NumberSequencesMode.PRIMES,          'maxIndex' : 1000,  'number' : '1560', 'expectedOutput' : PositionOfSequenceOutput('15601',
+          1818, 1)},
     ];
 
     for (var elem in _inputsToExpected) {
       test('sequence: ${elem['sequence']}, number: ${elem['start']}', () {
-        PositionOfSequenceOutput _actual = getFirstPositionOfSequence(elem['sequence'] as NumberSequencesMode, elem['number'] as String?, elem['maxIndex'] as int);
+        PositionOfSequenceOutput _actual = numberSequencesGetFirstPositionOfSequence(elem['sequence'] as NumberSequencesMode, elem['number'] as String?, elem['maxIndex'] as int);
         expect(_actual.number, (elem['expectedOutput'] as PositionOfSequenceOutput).number);
         expect(_actual.positionSequence, (elem['expectedOutput'] as PositionOfSequenceOutput).positionSequence);
         expect(_actual.positionDigits, (elem['expectedOutput'] as PositionOfSequenceOutput).positionDigits);
